pyproject.toml
[project]
name = "oilfield-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "django-ninja>=1.5.1",
    "uvicorn>=0.40.0",
]
